chore: codebase cleanup and agent optimization
- Fix all Clippy warnings (unused imports, dead code, derivable impls)
- Fix all Svelte/TS warnings (a11y roles, label associations)
- Rename package from d-developmentnethercore-projectframesmith to framesmith
- Update product name and identifier in Tauri config
- Fix silent error in list_characters (propagate instead of swallow)
- Add overflow protection with saturating_add in frame arithmetic
- Update move/state terminology in documentation
- Change zx-fspack.md status from Draft to Active
- Add Code Quality Standards section to CLAUDE.md
- Remove dead CharacterAssets struct and default_assets_version fn
- Derive Default for Pivot2, Movement, OnUse, OnHit, OnBlock structs

+++ src-tauri/src/codegen/zx_fspack_format.rs
@@ -271,7 +271,7 @@ mod tests {
 
     #[test]
     fn test_flags_reserved_is_zero() {
-        assert_eq!(FLAGS_RESERVED as u32, 0);
+        assert_eq!(FLAGS_RESERVED, 0);
     }
 
     #[test]
